Added A10 to exporters.md
Added a link in the 'Exporters' section for using the A10 exporter for older version of Thunder software.
Added a link in the 'metrics' section for using the native support in A10 Thunder. NOTE: This just links to the A10 main webpage currently. If we add something specific for Prometheus, I'll come back and change this.